fix: standardize decimals column to INTEGER across all tables
Decimals represents decimal places (typically 0-18), not a large
number. INTEGER is the appropriate type for both CIP-26 and CIP-68.

+++ extensions/asset-store/src/main/resources/db/store/mysql/V0_1500_1__init.sql
@@ -37,7 +37,7 @@ CREATE TABLE metadata_reference_nft (
     description    TEXT         NOT NULL,
     ticker         TEXT,
     url            TEXT,
-    decimals       BIGINT,
+    decimals       INTEGER,
     logo           LONGTEXT,
     version        BIGINT       NOT NULL,
     datum          LONGTEXT     NOT NULL,
